move unit tests to karma webpack
