+++ .gitignore
@@ -29,3 +29,6 @@ node_modules
 
 # Bower directory
 /vendor/
+
+# Webpack bundle- built on the fly with npm start
+dev/bundle.js
