+++ package.json
@@ -2,7 +2,7 @@
   "name": "rbx_ui_components",
   "version": "0.1.0",
   "scripts": {
-    "test": "echo 'NO TESTS'",
+    "test": "grunt test",
     "start": "webpack-dev-server --content-base dev --port 7000 --history-api-fallback"
   },
   "repository": {
