--- tests/js/unit/simple-unit-test.spec.js
@@ -1,9 +1,0 @@
-/*
- *  To Be Deleted once the first component tests are ready
- *
- */
-describe('Simple Unit Test', function () {
-    it ('should pass', function () {
-        expect(1).toEqual(1);
-    });
-});
